fix `UART_OPEN()` baud rate calculation
`UART_Open()` wasn't rounding up, which led to some baudrates being way off when fsys is 16mhz.

For example, the spec sheet says that a target baud rate of 115200 is supposed to have an actual baud rate of 111111.111, error rate of >3%. However, without rounding, the the actual baud rate is 125000, an error rate of 8%.

The reason this matters is because some serial interfaces can do the actual rates quoted in the spec sheet, and if they adjust their serial rate to match the expected rate, it will be way off.